+++ docs/Whats-New.md
@@ -315,7 +315,7 @@ New:
 - Unlimited skirmish colors (by Morton)
 - Example custom locomotor that circles around the target (by Kerbiter, CCHyper, with help from Otamaa; based on earlier experiment by CnCVK)
 - Vehicle voxel turret shadows & body multi-section shadows (by TwinkleStar)
-- `TiltsWhenCrushes` customization and wall crush slowdown toggle (by Starkku)
+- Crushing tilt and slowdown customization (by Starkku)
 
 Vanilla fixes:
 - Allow AI to repair structures built from base nodes/trigger action 125/SW delivery in single player missions (by Trsdy)